// src/app/core/models/user.model.ts
  export interface UsersModel extends Pagination {
    status: string;
    data: Users[];
    totalCount: number;
    filterText: string;
    limit: number;
    sortBy: string;
    sortByDirection: string
  }

  export interface UsersPaginationModel extends Pagination {
    docs: UsersModel[];
    filterText: string;
    sortBy: string;
    sortByDirection: string;
    status: string;
    title:string;
  }

  export interface Users {
    role: any;
    designation: any;
    _id: string;
    id:string;
    slug: string;
    name: string;
    phone_number: [];
    qualification: string;
    email: string;
    status: number;
    source: string;
    createdAt?: Date;
    updatedAt?: Date;
    type?: any;
    password: string;
    mobile?:String;
    joiningDate:Date,
    Active: any;
    menuItems: [];
    last_name:any;
    adminId: string;
    adminEmail: string;
    adminName: string;
    companyId:string;
    company:string;
    domain:string;
    users: number;
    courses: number;
    isLogin:boolean;
    phone:{
			countryCode:String,
			dialCode:String,
			e164Number:String,
			internationalNumber:String,
			nationalNumber:String,
			number:String,


		},
    avatar:string;
    filename:string
    education:string;
    attemptCalculation: number;

  }